+++ lib/async/http/server.rb
@@ -7,7 +7,6 @@
 require "async"
 require "io/endpoint"
 require "protocol/http/middleware"
-require "traces/provider"
 
 require_relative "protocol"
 
@@ -68,43 +67,6 @@ module Async
 					
 					# Wait for all children to finish:
 					task.children.each(&:wait)
-				end
-			end
-			
-			Traces::Provider(self) do
-				def call(request)
-					if trace_parent = request.headers["traceparent"]
-						Traces.trace_context = Traces::Context.parse(trace_parent.join, request.headers["tracestate"], remote: true)
-					end
-					
-					attributes = {
-						'http.version': request.version,
-						'http.method': request.method,
-						'http.authority': request.authority,
-						'http.scheme': request.scheme,
-						'http.path': request.path,
-						'http.user_agent': request.headers["user-agent"],
-					}
-					
-					if length = request.body&.length
-						attributes["http.request.length"] = length
-					end
-					
-					if protocol = request.protocol
-						attributes["http.protocol"] = protocol
-					end
-					
-					Traces.trace("async.http.server.call", attributes: attributes) do |span|
-						super.tap do |response|
-							if status = response&.status
-								span["http.status_code"] = status
-							end
-							
-							if length = response&.body&.length
-								span["http.response.length"] = length
-							end
-						end
-					end
 				end
 			end
 		end