+++ lib/traces/provider/async/http.rb
@@ -1,0 +1,2 @@
+require_relative "http/client"
+require_relative "http/server"